% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_outlier_value.R
\name{calculate_outlier_value}
\alias{calculate_outlier_value}
\title{Identify Outliers}
\usage{
calculate_outlier_value(x, tail, method = "iqr", na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{tail}{takes "upper" or "lower", determining which side of the distribution to calculate an outlier identification value.}

\item{method}{"iqr", the default, adds 1.5 times the interquartile range to the 75th percentile for the upper value or subtracts 1.5 times the interquartile range from the 25th percentile for the lower value. "sd" uses the mean plus or minus 3 times the standard deviation of the distribution.}

\item{na.rm}{a logical value indicating whether \code{NA} values should be stripped before the computation proceeds.}
}
\value{
A numeric vector of length one.
}
\description{
This function uses standard statistical methods to calculate the value the identifies outliers above or below it on a distribution.
}
\examples{
vec <- rnorm(1000, mean = 0, sd = 1)
calculate_outlier_value(vec, tail = "upper")
calculate_outlier_value(vec, tail = "lower", method = "sd")

}
