% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_lm2.R
\name{geom_lm2}
\alias{geom_lm2}
\title{A Convenient \code{ggplot2} Function for Linear Regression Plotting}
\usage{
geom_lm2(method = "lm", se = FALSE, formula = y ~ x, ...)
}
\arguments{
\item{method}{Smoothing method (function) to use, accepts either \code{NULL} or a character vector. Defaults to \code{stats::lm()}, unlike \code{ggplot2::geom_smooth()}.}

\item{se}{Display confidence interval around smooth? (\code{FALSE} by default, unlike \code{ggplot2::geom_smooth()}, see\code{ level} to control.)}

\item{formula}{Formula to use in smoothing function, eg. \code{y ~ x}, \code{y ~ poly(x, 2)}, \code{y ~ log(x)}. \code{NULL} by default, in which case \code{method = NULL} implies \code{formula = y ~ x} when there are fewer than 1,000 observations and \code{formula = y ~ s(x, bs = "cs")} otherwise. Unlike \code{ggplot2::geom_smooth()}, set by default so warning is not displayed.}

\item{...}{Additional arguments passed to \code{ggplot2::geom_smooth()}.}
}
\description{
This function is an exact duplication of \code{\link[ggplot2:geom_smooth]{ggplot2::geom_smooth()}} with defaults more appropriate for statistics and econometrics. By default, \code{\link[=geom_lm]{geom_lm()}} plots with \link{method = "lm"} for a linear trend line and with \link{se = FALSE} because plotting standard errors around a linear trend confuses the viewer more than it demonstrates statistical uncertainty.
}
\examples{
ggplot(ggplot2::mpg, aes(displ, hwy)) +
  ggplot2::geom_point() +
  geom_lm()
}
