% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{scatter}
\alias{scatter}
\title{Title}
\usage{
scatter(df, y, x)
}
\arguments{
\item{df}{Default dataset to use for plot.}

\item{y}{Values scattered on y-axis and on left-hand side of regression formula.}

\item{x}{Values scattered on x-axis and on right-hand side of regression formula.}
}
\value{
A \code{ggplot2} object
}
\description{
Title
}
\examples{
scatter(mtcars, hp, mpg)
}
