% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_lm.R
\name{geom_lm}
\alias{geom_lm}
\title{Plotting a linear trend}
\usage{
geom_lm(
  mapping = NULL,
  data = NULL,
  stat = "smooth",
  position = "identity",
  ...,
  method = "lm",
  formula = y ~ x,
  se = FALSE,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or \code{\link[=aes_]{aes_()}}. If specified and \link{inherit.aes = TRUE} (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping}

\item{data}{@param data The data to be displayed in this layer. There are three options:
If \code{NULL}, the default, the data is inherited from the plot data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot data. All objects will be fortified to produce a data frame. See \code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument, the plot data. The return value must be a \code{data.frame}, and will be used as the layer data. A \code{function} can be created from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{Used to override the default connection between \code{geom_smooth()} and \code{stat_smooth()}.}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{...}{Other arguments passed on to \code{ggplot2::layer()}. These are often aesthetics.}

\item{method}{Smoothing method (function) to use, accepts either \code{NULL} or a character vector. Defaults to \code{stats::lm()}, unlike \code{ggplot2::geom_smooth()}.}

\item{formula}{Formula to use in smoothing function, eg. \code{y ~ x}, \code{y ~ poly(x, 2)}, \code{y ~ log(x)}. \code{NULL} by default, in which case \code{method = NULL} implies \code{formula = y ~ x} when there are fewer than 1,000 observations and \code{formula = y ~ s(x, bs = "cs")} otherwise.}

\item{se}{Display confidence interval around smooth? (\code{FALSE} by default, see\code{ level} to control.)}

\item{na.rm}{IF \code{FALSE}, the default, missing values are removed with a warning. If \code{TRUE} missing values are silently removed.}

\item{orientation}{The orientation of the layer. The default (\code{NA}) automatically determines the orientation from the aesthetic mapping.}

\item{show.legend}{logical. Should this layer be included in the legends. \code{NA}, the default, includes if any aesthetics are mapped.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather than combining with them.}
}
\value{

}
\description{
This function is an exact duplication of \code{\link[ggplot2:geom_smooth]{ggplot2::geom_smooth()}} with defaults more appropriate for statistics and econometrics. By default, \code{\link[=geom_lm]{geom_lm()}} plots with \link{method = "lm"} for a linear trend line and with \link{se = FALSE} because plotting standard errors around a linear trend confuses the viewer more than it demonstrates statistical uncertainty.
}
\examples{
ggplot(ggplot2::mpg, aes(displ, hwy)) +
  ggplot2::geom_point() +
  geom_lm()
}
